Updated metadata from level 1 meeting. Uncommented date to account for DateTime fix. Moved level1.py to better reflect schema structure.
