+++ romancal/datamodels/schemas/image_files/image.schema.yaml
@@ -31,9 +31,9 @@ allOf:
             sub_category:
               title: Program sub-category
               type: string
-#            science_category:
-#              title: Science category assigned during TAC process
-#              type: string
+            science_category:
+              title: Science category assigned during TAC process
+              type: string
             continuation_id:
               title: Continuation of previous program
               type: integer
@@ -84,9 +84,6 @@ allOf:
             visit_group:
               title: Visit group identifier
               type: string
-            sequence_id:
-              title: Parallel sequence identifier
-              type: string
             activity_id:
               title: Activity identifier
               type: string
@@ -102,10 +99,22 @@ allOf:
             observation_folder:
               title: Name of the APT observation folder
               type: string
+            ma_table_name:
+              title: Identifier for the multi-accumulation table used
+              type: string
         ephemeris:
           title: Ephemeris data information
           type: object
           properties:
+            earth_angle:
+              title: "Earth Angle [radians]"
+              type: number
+            moon_angle:
+              title: "Moon Angle [radians]"
+              type: number
+            sun_angle:
+              title: "Sun Angle [radians]"
+              type: number
             type:
               title: Definitive or Predicted
               type: string
@@ -148,9 +157,9 @@ allOf:
             start_time:
               title: UTC visit start time
               type: string
-#            end_time:
-#              title: UTC visit end time
-#              type: string
+            end_time:
+              title: UTC visit end time
+              type: string
             status:
               title: Status of a visit
               type: string
@@ -190,9 +199,6 @@ allOf:
           type: object
           properties:
             name:
-              title: PRD science aperture used
-              type: string
-            pps_name:
               title: PRD science aperture used
               type: string
             position_angle:
@@ -344,75 +350,75 @@ allOf:
 #            gs_order:
 #              title: index of guide star within list of selected guide stars
 #              type: integer
-#            gs_start_time:
-#              title: UTC time when guide star activity started
-#              type: string
-#            gs_stop_time:
-#              title: UTC time when guide star activity completed
-#              type: string
-#            gs_id:
-#              title: guide star identifier
-#              type: string
-#            gs_ra:
-#              title: guide star right ascension
-#              type: number
-#            gs_dec:
-#              title: guide star declination
-#              type: number
-#            gs_ura:
-#              title: guide star right ascension uncertainty
-#              type: number
-#            gs_udec:
-#              title: guide star declination uncertainty
-#              type: number
-#            gs_mag:
-#              title: guide star magnitude in FGS detector
-#              type: number
-#            gs_umag:
-#              title: guide star magnitude uncertainty
-#              type: number
-#            gs_pcs_mode:
-#              title: Pointing Control System mode
-#              type: string
-#            gs_function_start_time:
-#              title: Observatory UTC time at guider function start
-#              type: string
-#            gs_function_end_time:
-#              title: Observatory UTC time at guider function end
-#              type: string
-#            data_start:
-#              title: MJD start time of guider data within this file
-#              type: number
-#            data_end:
-#              title: MJD end time of guider data within this file
-#              type: number
-#            gs_acq_exec_stat:
-#              title: Guide star acquisition execution status
-#              type: string
+            gs_start_time:
+              title: UTC time when guide star activity started
+              type: string
+            gs_stop_time:
+              title: UTC time when guide star activity completed
+              type: string
+            gs_id:
+              title: guide star identifier
+              type: string
+            gs_ra:
+              title: guide star right ascension
+              type: number
+            gs_dec:
+              title: guide star declination
+              type: number
+            gs_ura:
+              title: guide star right ascension uncertainty
+              type: number
+            gs_udec:
+              title: guide star declination uncertainty
+              type: number
+            gs_mag:
+              title: guide star magnitude in FGS detector
+              type: number
+            gs_umag:
+              title: guide star magnitude uncertainty
+              type: number
+            gs_pcs_mode:
+              title: Pointing Control System mode
+              type: string
+            gs_function_start_time:
+              title: Observatory UTC time at guider function start
+              type: string
+            gs_function_end_time:
+              title: Observatory UTC time at guider function end
+              type: string
+            data_start:
+              title: MJD start time of guider data within this file
+              type: number
+            data_end:
+              title: MJD end time of guider data within this file
+              type: number
+            gs_acq_exec_stat:
+              title: Guide star acquisition execution status
+              type: string
             gs_ctd_x:
               title: "[arcsec] guide star centroid x position in FGS ideal frame"
               type: number
             gs_ctd_y:
               title: "[arcsec] guide star centroid y position in FGS ideal frame"
               type: number
-#            gs_ctd_ux:
-#              title: uncertainty in the x position of the centroid
-#              type: number
-#            gs_ctd_uy:
-#              title: uncertainty in the y position of the centroid
-#              type: number
-#            gs_epoch:
-#              title: Epoch of guide star coordinates
-#              type: string
-#            gs_mura:
-#              title: Guide star IRCS right ascension proper motion
-#              type: number
-#            gs_mudec:
-#              title: Guide star IRCS declination proper motion
-#              type: number
-#            gs_para:
-#              title: Guide star annual parallax
-#              type: number
+            gs_ctd_ux:
+              title: uncertainty in the x position of the centroid
+              type: number
+            gs_ctd_uy:
+              title: uncertainty in the y position of the centroid
+              type: number
+            gs_epoch:
+              title: Epoch of guide star coordinates
+              type: string
+            gs_mura:
+              title: Guide star IRCS right ascension proper motion
+              type: number
+            gs_mudec:
+              title: Guide star IRCS declination proper motion
+              type: number
+            gs_para:
+              title: Guide star annual parallax
+              type: number
             gs_window_xstart:
               title: Guide window x position
               type: number
